+++ src/Application/GenerateApplication.php
@@ -3,6 +3,7 @@
 namespace SatisGen\Application;
 
 use SatisGen\Command\GenerateCommand;
+use SatisGen\Config\ConfigInputReader;
 use Symfony\Component\Console\Application;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Filesystem\Filesystem;
@@ -11,10 +12,12 @@ use Symfony\Component\Filesystem\Filesystem;
 class GenerateApplication extends Application
 {
     private $filesystem;
+    private $config;
     
-    public function __construct(Filesystem $filesystem) 
+    public function __construct(Filesystem $filesystem, ConfigInputReader $config) 
     {
         $this->filesystem = $filesystem;
+        $this->config = $config;
         parent::__construct('Satis Generator', '0.0.1');
     }
     
@@ -26,7 +29,7 @@ class GenerateApplication extends Application
     protected function getDefaultCommands()
     {
         $defaultCommands = parent::getDefaultCommands();
-        $defaultCommands[] = new GenerateCommand($this->filesystem);
+        $defaultCommands[] = new GenerateCommand($this->filesystem, $this->config);
         return $defaultCommands;
     }