+++ tests/SatisGenTest.php
@@ -3,7 +3,9 @@
 namespace SatisGen\Tests;
 
 use SatisGen\Application\GenerateApplication;
+use SatisGen\Config\ConfigInputReader;
 use Symfony\Component\Filesystem\Filesystem;
+
 
 abstract class SatisGenTest extends \PHPUnit_Framework_TestCase
 {
@@ -16,8 +18,11 @@ abstract class SatisGenTest extends \PHPUnit_Framework_TestCase
         // Filesystem
         $this->filesystem = new Filesystem();
 
+        // ConfigInputReader
+        $this->configInputReader = new ConfigInputReader();
+
         // GenerateApplication
-        $application = new GenerateApplication($this->filesystem);
+        $application = new GenerateApplication($this->filesystem, $this->configInputReader);
         $application->setAutoExit(false);
         $this->application = $application;
     }